fix tests error caused by same package and file name move all tests to tests package refactor tests files to start with test prefix
