+++ api/service/tests/__init__.py
@@ -1,4 +1,4 @@
-from .telemetry import TelemetryFileParsing
-from .asset import AssetFileParsing
-from .api import AssetTests
-from .api import MissionTests
+from .test_telemetry import TelemetryFileParsing
+from .test_asset import AssetFileParsing
+from .test_api import AssetTests
+from .test_api import MissionTests
