+++ api/service/tests/test_asset.py
