+++ api/service/tests/test_telemetry.py
